runsvdir.c: new option "-L fifo" to log stderr to the named pipe fifo
The option -L /path/to/fifo tells runsvdir to open /path/to/fifo, which
shall be a named pipe created by mkfifo(1) or similar. If fifo does not
exist, is not writable, or not readable, runsvdir tries again later.

After fifo is opened for writing and reading, runsvdir writes data that
is available on stderr to it.  If the fifo buffer is full because the
data is not read, runsvdir re-opens fifo to empty the buffer, and tries
to write a warning to the fifo indicating that the log is incomplete.
Emptying the buffer doesn't work though, if there is a reader that is
not reading. In this case the newer data is lost instead of the older.

To consume the log, "cat fifo" can be used, better yet, a service like

 #!/bin/sh
 exec < fifo
 exec svlogd -t -l200 -b256 .

Also, runsvdir now uses sgetopt for option parsing, and keeps one less
file descriptor open for the readproctitle log.