chpst.c: add new option -F: fghack using unshare().
chpst with option -F runs prog in new mount and pid namespaces as pid 2.
It provides a pid 1 that whenever prog or a detached child of prog exits
checks /proc whether it's alone or there's still prog or children of it
running. If it finds itself alone in the pid namespace, it exits. While
running, chpst -F relays signals it receives to the pid 1 it provides,
which in turn relays the signal to all children in the pid namespace.